delete comment. move IntToObjArray to new file. small initialCapacity. loadFactor
