+++ src/com/esotericsoftware/kryo/util/ArrayClassResolver.java
@@ -26,34 +26,28 @@ import com.esotericsoftware.kryo.Registration;
 import com.esotericsoftware.kryo.io.Input;
 import com.esotericsoftware.kryo.io.Output;
 
-import java.lang.reflect.Array;
-
 import static com.esotericsoftware.kryo.util.Util.*;
 import static com.esotericsoftware.minlog.Log.*;
 
 /** This is enhanced resolver from {@link DefaultClassResolver} for fast deserialization of collections.
  *
+ * @author lifeinwild1@gmail.com
  * @apiNote In terms of functionality, {@link ArrayClassResolver} is completely equivalent to {@link DefaultClassResolver}. So
  *          output binary of {@link ArrayClassResolver} is equivalent to that of {@link DefaultClassResolver}.
- *
  * @implNote You can specify the mappings between class and ID by {@link Kryo#register(Class, int)}, But don't specify huge ID
  *           like 20000000 because this resolver uses array internally. This resolver internally reconstructs
  *           {@link IntToObjArray#array} when the mappings are updated. Therefore, it is not suitable in terms of performance if
  *           the mappings are added frequently at peaktime of application. Use the {@link Pool}.
- * @see <a href="https://github.com/EsotericSoftware/kryo#pooling">Pool</a>
- *
- * @author lifeinwild1@gmail.com */
-public final class ArrayClassResolver implements ClassResolver {
-	protected Kryo kryo;
-
+ * @see <a href="https://github.com/EsotericSoftware/kryo#pooling">Pool</a> */
+public class ArrayClassResolver implements ClassResolver {
 	protected final IdentityMap<Class, Registration> classToRegistration = new IdentityMap<>();
-
+	private final IntToObjArray<Registration> idToRegistrationTmp = new IntToObjArray<>(Registration.class);
+	protected Kryo kryo;
 	protected IdentityObjectIntMap<Class> classToNameId;
 	protected IntToObjArray<Class> nameIdToClass;
 	protected ObjectMap<String, Class> nameToClass;
 	protected int nextNameId;
 
-	/** TODO: Map support */
 	private int memoizedClassId = -1;
 	private Registration memoizedClassIdValue;
 	private Class memoizedClass;
@@ -137,7 +131,7 @@ public final class ArrayClassResolver implements ClassResolver {
 
 	protected Registration readName (Input input) {
 		int nameId = input.readVarInt(true);
-		if (nameIdToClass == null) nameIdToClass = new IntToObjArray<>(Class.class);
+		if (nameIdToClass == null) nameIdToClass = new IntToObjArray<>(Class.class, 20);
 		Class type = nameIdToClass.array[nameId];
 		if (type == null) {
 			// Only read the class name the first time encountered in object graph.
@@ -213,50 +207,6 @@ public final class ArrayClassResolver implements ClassResolver {
 
 		return registration;
 	}
-
-	/** replacement of {@link IntMap} for faster lookup. */
-	private static class IntToObjArray<E> {
-		public E[] array;
-		private final Class<E> valueType;
-		private final int initialCapacity;
-
-		/** @param valueType for generic type array {@link #array} */
-		IntToObjArray (Class<E> valueType, int initialCapacity) {
-			this.valueType = valueType;
-			array = (E[])Array.newInstance(valueType, initialCapacity);
-			this.initialCapacity = initialCapacity;
-		}
-
-		IntToObjArray (Class<E> valueType) {
-			this(valueType, 1000);
-		}
-
-		public void clear () {
-			array = (E[])Array.newInstance(valueType, initialCapacity);
-		}
-
-		public E remove (int classid) {
-			if (classid >= array.length)
-				return null;
-			E r = array[classid];
-			array[classid] = null;
-			return r;
-		}
-
-		public E put (int classid, E v) {
-			if (classid >= array.length) {
-				E[] next = (E[])Array.newInstance(valueType, (int)(array.length * 1.1));
-				System.arraycopy(array, 0, next, 0, array.length);
-				array = next;
-			}
-
-			E r = array[classid];
-			array[classid] = v;
-			return r;
-		}
-	}
-
-	private IntToObjArray<Registration> idToRegistrationTmp = new IntToObjArray(Registration.class);
 
 	public Registration unregister (int classID) {
 		Registration registration = idToRegistrationTmp.remove(classID);
